Go to given url instead of Pocket app.
Going to pocket web is a bad experience when the user is not logged in.
What would happen is that a non-logged in user will be redirected to the login page,
and on login be redirected to the home page instead of to the pocket link as given in the email.
Hence better UX is to just go to the resolved link.